09_LARGE_LANGUAGE_MODELS/04_LLM_IN_PRODUCTION/01_LECTURE/streamlit_LLM_integration/requirements.txt
streamlit
langchain_mistralai
python-dotenv